# contracts/admin/openapi.json
{
  "components": {
    "schemas": {
...
      "BindBody": {
        "additionalProperties": false,
        "properties": {
          "state": {
            "maxLength": 512,
            "minLength": 16,
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state"
        ],
        "title": "BindBody",
        "type": "object"
      },
...
      "CallbackBody": {
        "additionalProperties": false,
        "properties": {
          "connection_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connection Id"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "email_verified": {
            "title": "Email Verified",
            "type": "boolean"
          },
          "identity_type": {
            "title": "Identity Type",
            "type": "string"
          },
          "issuer": {
            "title": "Issuer",
            "type": "string"
          },
          "organization_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Id"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "state": {
            "maxLength": 512,
            "minLength": 16,
            "title": "State",
            "type": "string"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          }
        },
        "required": [
          "state",
          "provider",
          "issuer",
          "subject",
          "email",
          "email_verified",
          "identity_type"
        ],
        "title": "CallbackBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StartAuthBody": {
        "additionalProperties": false,
        "properties": {
          "requested_connection_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Connection Id"
          },
          "requested_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Provider"
          },
          "return_url": {
            "anyOf": [
              {
                "maxLength": 1024,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return Url"
          }
        },
        "title": "StartAuthBody",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Administrative API for FrostGate management console",
    "title": "FrostGate Admin Gateway",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/identity/invitations/{invitation_id}/bind": {
      "post": {
        "operationId": "bind_identity_invitations__invitation_id__bind_post",
        "parameters": [
          {
            "in": "path",
            "name": "invitation_id",
            "required": true,
            "schema": {
              "title": "Invitation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": true,
            "schema": {
              "title": "X-Tenant-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bind",
        "tags": [
          "identity-enforcement"
        ]
      }
    },
    "/identity/invitations/{invitation_id}/callback": {
      "post": {
        "operationId": "callback_identity_invitations__invitation_id__callback_post",
        "parameters": [
          {
            "in": "path",
            "name": "invitation_id",
            "required": true,
            "schema": {
              "title": "Invitation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": true,
            "schema": {
              "title": "X-Tenant-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CallbackBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Callback Identity Invitations  Invitation Id  Callback Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Callback",
        "tags": [
          "identity-enforcement"
        ]
      }
    },
    "/identity/invitations/{invitation_id}/requirements": {
      "get": {
        "operationId": "invitation_requirements_identity_invitations__invitation_id__requirements_get",
        "parameters": [
          {
            "in": "path",
            "name": "invitation_id",
            "required": true,
            "schema": {
              "title": "Invitation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": true,
            "schema": {
              "title": "X-Tenant-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Invitation Requirements Identity Invitations  Invitation Id  Requirements Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Invitation Requirements",
        "tags": [
          "identity-enforcement"
        ]
      }
    },
    "/identity/invitations/{invitation_id}/start-auth": {
      "post": {
        "operationId": "start_auth_identity_invitations__invitation_id__start_auth_post",
        "parameters": [
          {
            "in": "path",
            "name": "invitation_id",
            "required": true,
            "schema": {
              "title": "Invitation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": true,
            "schema": {
              "title": "X-Tenant-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartAuthBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Start Auth Identity Invitations  Invitation Id  Start Auth Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Auth",
        "tags": [
          "identity-enforcement"
        ]
      }
    },
...
  }
}